+++ build-wasm.nix
@@ -21,6 +21,7 @@ let
       chmod -R 777 .
 
       test -e LICENSE && cp LICENSE $out/tree-sitter-${id}.LICENSE
+      test -e LICENSE.md && cp LICENSE.md $out/tree-sitter-${id}.LICENSE.md
       test -e COPYING.txt && cp COPYING.txt $out/tree-sitter-${id}.COPYING.txt
 
       ${preBuild}
@@ -38,21 +39,30 @@ in
 pkgs.symlinkJoin {
   name = "tree-sitter-wasm-builds-combined";
   paths = with sources; [
+    (grammar "bash" tree-sitter-bash)
     (grammar "c" tree-sitter-c)
+    (grammar "clojure" tree-sitter-clojure)
     (grammar "css" tree-sitter-css)
+    (grammar "fennel" tree-sitter-fennel)
+    (grammar "go" tree-sitter-go)
+    (grammar "haskell" tree-sitter-haskell)
     (grammar "html" tree-sitter-html)
     (grammar "java" tree-sitter-java)
     (grammar "javascript" tree-sitter-javascript)
     (grammar "json" tree-sitter-json)
-    (grammar "python" tree-sitter-python)
-    (grammar "clojure" tree-sitter-clojure)
-    (grammar "fennel" tree-sitter-fennel)
+    (grammar "lua" tree-sitter-lua)
     (grammar "markdown" tree-sitter-markdown)
     (grammar "nix" tree-sitter-nix)
+    (mkTreeSitterGrammar "ocaml" tree-sitter-ocaml
+      # contains multiple grammars: ocaml (.ml) ; interface (.mli)
+      { preBuild = "cd ocaml"; })
+    (grammar "php" tree-sitter-php)
+    (grammar "python" tree-sitter-python)
     (grammar "scss" tree-sitter-scss)
-    # (grammar "yaml" "https://github.com/ikatyang/tree-sitter-yaml") FIXME: build fails
+    (grammar "toml" tree-sitter-toml)
     (mkTreeSitterGrammar "typescript" tree-sitter-typescript
-      # repository contains multiple grammars in sub-directories: TypeScript and JSX
+      # contains multiple grammars: TypeScript and JSX
       { preBuild = "cd typescript"; })
+    (grammar "yaml" tree-sitter-yaml)
   ];
 }