+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is roughly based on [Keep a Changelog](https://keepachangelog.com/).
 
 - pin grammar dependencies with, managed with niv
 - update web-tree-sitter to 0.19.4, tree-sitter-cli to 0.19.3, emscripten to 2.0.27
+- add more languages: bash, go, haskell, lua, ocaml, php, toml, yaml
 
 ## [0.3.0]